Rename Examples.TimeSystem in Examples
