+++ Docs/Examples/Examples.csproj
@@ -28,10 +28,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\Source\Testably.Abstractions\Testably.Abstractions.csproj" />
-    <ProjectReference
-      Include="..\..\..\Source\Testably.Abstractions.Extensions\Testably.Abstractions.Extensions.csproj" />
-    <ProjectReference Include="..\..\..\Source\Testably.Abstractions.Testing\Testably.Abstractions.Testing.csproj" />
+    <ProjectReference Include="..\..\Source\Testably.Abstractions\Testably.Abstractions.csproj" />
+    <ProjectReference Include="..\..\Source\Testably.Abstractions.Extensions\Testably.Abstractions.Extensions.csproj" />
+    <ProjectReference Include="..\..\Source\Testably.Abstractions.Testing\Testably.Abstractions.Testing.csproj" />
   </ItemGroup>
 
 </Project>
